requirements.txt
nanpy==0.9.6
pkg-resources==0.0.0
pyserial==3.4
pyzmq==17.0.0
zmq==0.0.0
